requirements.txt
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cryptography==41.0.2
faiss-cpu==1.7.4
gensim==4.3.1
joblib==1.3.1
nltk==3.8.1
numpy==1.25.1
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==10.0.0
pycparser==2.21
PyPDF2==3.0.1
regex==2023.6.3
scikit-learn==1.3.0
scipy==1.11.1
smart-open==6.3.0
threadpoolctl==3.1.0
tqdm==4.65.0
Wand==0.6.11
